# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cadastro Service API",
    "version": "2.0.0",
    "description": "API central de autenticação e SSO da equipe de Cadastro."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Ambiente local"
    }
  ],
  "tags": [
    { "name": "Saúde" },
    { "name": "Autenticação" },
    { "name": "Admin" }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["Saúde"],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Aplicação disponível",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Cadastro de usuário local",
        "description": "Cria um usuário com senha forte e retorna tokens de sessão.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" },
              "example": {
                "email": "usuario@example.com",
                "password": "Senha@Forte123",
                "confirmPassword": "Senha@Forte123",
                "name": "Nome do Usuário",
                "dateOfBirth": "1995-05-20",
                "role": "client"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário cadastrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou e-mail já utilizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Login com e-mail e senha",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sessão criada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/google": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Login com token do Google",
        "description": "Sincroniza o usuário a partir do Google Identity e devolve os tokens locais.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GoogleLoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sessão criada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Token do Google inválido",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Rotaciona o refresh token",
        "description": "Recebe um refresh token válido, revoga-o e retorna novo par access/refresh.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sessão renovada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          },
          "400": {
            "description": "Payload inválido",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Refresh token inválido ou expirado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Revoga um refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Refresh token revogado" },
          "400": {
            "description": "Payload inválido",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Autenticação"],
        "summary": "Perfil autenticado",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Dados do usuário autenticado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponse" }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/validate": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Validação de access token",
        "description": "Permite que outros serviços validem tokens emitidos pelo cadastro.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ValidateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token válido",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenValidationResponse" }
              }
            }
          },
          "400": {
            "description": "Payload inválido",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Token inválido",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenInvalidResponse" }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenInvalidResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "Lista usuários",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsersResponse" }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Usuário sem permissão",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/users/{id}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Perfil detalhado",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil e permissões do usuário",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponse" }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Usuário sem permissão",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "PublicUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": ["string", "null"], "format": "email" },
          "name": { "type": "string" },
          "dateOfBirth": { "type": ["string", "null"], "format": "date" },
          "googleId": { "type": ["string", "null"] },
          "role": { "type": "string", "enum": ["client", "admin"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "SessionResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string", "description": "JWT válido por 15 minutos" },
          "refreshToken": { "type": "string", "description": "Token opaco válido por 30 dias" },
          "user": { "$ref": "#/components/schemas/PublicUser" },
          "permissions": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/PublicUser" },
          "permissions": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "UsersResponse": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PublicUser" }
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "confirmPassword", "name"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password" },
          "confirmPassword": { "type": "string", "format": "password" },
          "name": { "type": "string" },
          "dateOfBirth": { "type": ["string", "null"], "format": "date" },
          "role": { "type": "string", "enum": ["client", "admin"], "default": "client" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password" }
        }
      },
      "GoogleLoginRequest": {
        "type": "object",
        "required": ["idToken"],
        "properties": {
          "idToken": { "type": "string" }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string" }
        }
      },
      "ValidateRequest": {
        "type": "object",
        "required": ["token"],
        "properties": {
          "token": { "type": "string" }
        }
      },
      "TokenValidationResponse": {
        "type": "object",
        "properties": {
          "valid": { "type": "boolean", "example": true },
          "user": { "$ref": "#/components/schemas/PublicUser" },
          "permissions": {
            "type": "array",
            "items": { "type": "string" }
          },
          "token": {
            "type": "object",
            "properties": {
              "subject": { "type": "string", "format": "uuid" },
              "email": { "type": ["string", "null"] },
              "role": { "type": "string", "enum": ["client", "admin"] }
            }
          }
        }
      },
      "TokenInvalidResponse": {
        "type": "object",
        "properties": {
          "valid": { "type": "boolean", "example": false },
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
      "ValidationErrorResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ErrorResponse" },
          {
            "type": "object",
            "properties": {
              "issues": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": { "type": "array", "items": { "type": "string" } },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        ]
      }
    }
  }
}